+++ hutool-core/src/test/java/cn/hutool/core/math/CalculatorTest.java
@@ -1,7 +1,6 @@
 package cn.hutool.core.math;
 
 import org.junit.Assert;
-import org.junit.Ignore;
 import org.junit.Test;
 
 public class CalculatorTest {
@@ -31,7 +30,7 @@ public class CalculatorTest {
 	}
 
 	@Test
-	@Ignore
+//	@Ignore
 	public void conversationTest5(){
 		// https://github.com/dromara/hutool/issues/1984
 		final double conversion = Calculator.conversion("((1/1) / (1/1) -1) * 100");