+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * 【db     】     Db增加query重载，可支持自定义PreparedStatement，从而支持游标（issue#I4JXWN@Gitee）
 * 【cache  】     CacheObj增加getExpiredTime等方法（issue#I4LE80@Gitee）
 * 【extra  】     Ftp增加backToPwd方法（issue#2004@Github）
+* 【core   】     CollStreamUtil修改集合中null处理问题（pr#478@Gitee）
 * 
 ### 🐞Bug修复
 * 【core   】     修复FileResource构造fileName参数无效问题（issue#1942@Github）